CHANGED: Trunk - Additional: Was made changes in script elastix-menuremove. Before this script given a Menu deleted this menu and all menu that was children of this. Now this script given a directory that store a bounch of xml files, delete the menu that represent those files

+++ framework/system/libs/paloSantoACL.class.php
@@ -1444,6 +1444,12 @@ INFO_AUTH_MODULO;
      ******************************************************************/
     function deleteResource($idresource)
     {
+        //validamos el id del recurso
+        if(!preg_match("/^[[:word:]]+$/",$idresource)){
+            $this->errMsg=_tr("Invalid Resource");
+            return false;
+        }
+        
         $this->errMsg = "";
         $query = "DELETE FROM acl_resource WHERE id = ?";
         $result = $this->_DB->genQuery($query,array($idresource));
